Refactor script module into package
